pyproject.toml
[tool.poetry]
name = "BI Tool"
version = "0.1.0"
description = ""
authors = ["graphicmist", "alex", "exhale"]
packages = [{ include = "core", from = "src" }]

[tool.poetry.dev-dependencies]
mypy = "*"
pytest = "*"
markdown = "*"
bump2version = "*"
isort = { extras = ["pyproject"], version = "*" }
black = "*"
coverage = { extras = ["toml"], version = "^5.5" }
pytest-cov = "^2.12.1"
requests-mock = "^1.9.3"

[tool.poetry.dependencies]
python = "^3.9"
ipdb = "*"
fire = "*"
numpy = "1.26.4"
sentencepiece = "0.1.97"
transformers = "^4.15.0"
fastapi = { version = "^0.100.0", optional = true }
fasttext-wheel = { version = "*", optional = true }
json-logging = { version = "^1.3.0", optional = true }
PyYAML = { version = "6.0", optional = true }
omegaconf = "^2.1.1"
uvicorn = { extras = ["standard"], version = "^0.24.0", optional = true }
ddtrace = { version = "^2.3.1", optional = true }
h5py = "^3.6.0"
pydantic = { version = "^2.6.4", extras = ["dotenv"] }
PyDrive2 = { version = "^1.10.0", optional = true }
s3fs = { version = "^2022.02.0", optional = true }
boto3 = { version = "*", optional = true }
dvc = { version = "^2.9.3", extras = ["s3"], optional = true }
requests = "^2.27.0"
hashids = "^1.3.1"
openai = "1.54.4"
langcodes = "3.3.0"
language-data = "1.1"
nltk = "^3.8.1"
pillow = "^9.5.0"
annoy = "^1.17.1"
jupyter = "^1.0.0"
selenium = "^4.9.1"
bs4 = "^0.0.1"
weaviate-client = "^3.25.3"
supabase = "^2.4.1"
watchfiles = "^0.21.0"
tiktoken = "^0.8.0"
sse-starlette = "^1.8.2"
slowapi = "^0.1.8"
pyjwt = "^2.8.0"
langchain-openai = "^0.2.9"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
grandalf = "^0.8"
pydantic-settings = "^2.2.1"
pandas = "^2.2.1"
typer = "^0.12.0"
redis = "^5.0.3"
loguru = "^0.7.2"
langchain-anthropic = "^0.3.0"
langsmith = "^0.1.144"
langchain-core = "^0.3.19"

[tool.poetry.extras]
api = ["fastapi", "ddtrace", "json-logging", "PyYAML", "uvicorn"]
training = ["dvc", "pydrive2", "boto3", "s3fs", "fasttext"]

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = "context.py"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

[requires]
python_version = "^3.11.0"

